Inter-cluster data locality
Implement a simple mechanism of data locality between clusters.
If a cluster has "compute" in the template name, then another
cluster with "storage" in the template name is looked for. If found,
the instances of the new cluster will be created on the same
hosts as the ones of the storage cluster, using the SameHost Nova
filter.

Change-Id: I1b88e57ea53dc55a12a8c13508aea9361099e9cc